A re-think of how to handle separators in order to support both header and footer or other emphasized (strong) separators.
Removed class inheritance and logic for previous/next row and replaced with adding a non-default border_type into the table elaboration logic.

Also split the table elaboration (of implicit separators) from the rendering portion so that it can be intercepted in the case where a user would want to modify the elaborated separators.